+++ projects/angular-tiptap-editor/src/lib/services/link.service.ts
@@ -1,9 +1,7 @@
 import { Injectable, signal } from "@angular/core";
 import type { Editor } from "@tiptap/core";
 
-@Injectable({
-    providedIn: "root",
-})
+@Injectable()
 export class LinkService {
 
     // ============================================